+++ lib/waterdrop.rb
@@ -3,12 +3,11 @@
 # External components
 # delegate should be removed because we don't need it, we just add it because of ruby-kafka
 %w[
-  karafka-core
   forwardable
-  rdkafka
   json
   zeitwerk
   securerandom
+  karafka-core
 ].each { |lib| require lib }
 
 # WaterDrop library
@@ -27,9 +26,3 @@ loader.inflector.inflect('waterdrop' => 'WaterDrop')
 loader.ignore("#{__dir__}/waterdrop/instrumentation/vendors/**/*.rb")
 loader.setup
 loader.eager_load
-
-# Rdkafka uses a single global callback for things. We bypass that by injecting a manager for
-# each callback type. Callback manager allows us to register more than one callback
-# @note Those managers are also used by Karafka for consumer related statistics
-Rdkafka::Config.statistics_callback = WaterDrop::Instrumentation.statistics_callbacks
-Rdkafka::Config.error_callback = WaterDrop::Instrumentation.error_callbacks